+++ Core/Abstractions/GroupId.cs
@@ -3,37 +3,42 @@ using System;
 namespace Unity.Cloud.Common
 {
     /// <summary>
-    /// This struct holds information about a permission.
+    /// This struct holds information about a group identifier.
     /// </summary>
-    public readonly struct Permission
+    public readonly struct GroupId
     {
         readonly string m_String;
 
         /// <summary>
-        /// Creates a <see cref="Permission"/> using a <see cref="string"/>.
+        /// Returns the value of an identifier representing an invalid group id
         /// </summary>
-        /// <param name="value">The string representing the permission</param>
-        public Permission(string value) => m_String = value;
+        public static readonly GroupId None = new(string.Empty);
 
         /// <summary>
-        /// Returns whether two <see cref="Permission"/> objects are equals.
+        /// Creates a <see cref="GroupId"/> using a <see cref="string"/>.
+        /// </summary>
+        /// <param name="value">The string representing the group identifier</param>
+        public GroupId(string value) => m_String = value;
+
+        /// <summary>
+        /// Returns whether two <see cref="GroupId"/> objects are equals.
         /// </summary>
         /// <param name="other"></param>
         /// <returns>
         /// <see langword="true"/> if both instances have the same values;
         /// <see langword="false"/> otherwise.
         /// </returns>
-        public bool Equals(Permission other) => m_String == other.m_String;
+        public bool Equals(GroupId other) => m_String == other.m_String;
 
         /// <summary>
-        /// Validates that  <paramref name="obj"/> is a <see cref="Permission"/> instance and that it has the same value as this instance.
+        /// Validates that  <paramref name="obj"/> is a <see cref="GroupId"/> instance and that it has the same value as this instance.
         /// </summary>
         /// <param name="obj">Compare the values with this instance.</param>
         /// <returns>
         /// <see langword="true"/> if both instances have the same values;
         /// <see langword="false"/> otherwise.
         /// </returns>
-        public override bool Equals(object obj) => obj is UserId other && Equals(other);
+        public override bool Equals(object obj) => obj is GroupId other && Equals(other);
 
         /// <summary>
         /// Computes a hash code for the object.
@@ -47,13 +52,13 @@ namespace Unity.Cloud.Common
         public override int GetHashCode() => m_String != null ? m_String.GetHashCode() : 0;
 
         /// <summary>
-        /// Get the string representation of this <see cref="Permission"/>.
+        /// Get the string representation of this <see cref="GroupId"/>.
         /// </summary>
         /// <returns>The string result.</returns>
         public override string ToString() => m_String;
 
         /// <summary>
-        /// Returns whether two <see cref="Permission"/> instances are equal.
+        /// Returns whether two <see cref="GroupId"/> instances are equal.
         /// </summary>
         /// <param name="left">First instance.</param>
         /// <param name="right">Second instance.</param>
@@ -61,10 +66,10 @@ namespace Unity.Cloud.Common
         /// <see langword="true"/> if the instances are equal;
         /// <see langword="false"/> otherwise.
         /// </returns>
-        public static bool operator ==(Permission left, Permission right) => left.Equals(right);
+        public static bool operator ==(GroupId left, GroupId right) => left.Equals(right);
 
         /// <summary>
-        /// Returns whether two <see cref="Permission"/> instances are different.
+        /// Returns whether two <see cref="GroupId"/> instances are different.
         /// </summary>
         /// <param name="left">First instance.</param>
         /// <param name="right">Second instance.</param>
@@ -72,13 +77,13 @@ namespace Unity.Cloud.Common
         /// <see langword="true"/> if the instances are different;
         /// <see langword="false"/> otherwise.
         /// </returns>
-        public static bool operator !=(Permission left, Permission right) => !left.Equals(right);
+        public static bool operator !=(GroupId left, GroupId right) => !left.Equals(right);
 
         /// <summary>
-        /// Explicitly cast a <see cref="Permission"/> to a <see cref="string"/>
+        /// Explicitly cast a <see cref="GroupId"/> to a <see cref="string"/>
         /// </summary>
-        /// <param name="permission">Object to cast</param>
+        /// <param name="uId">Object to cast</param>
         /// <returns>The resulting <see cref="string"/></returns>
-        public static explicit operator string(Permission permission) => permission.m_String;
+        public static explicit operator string(GroupId uId) => uId.m_String;
     }
 }